fix(indev): change zero_streak from uint8_t to uint32_t to prevent overflow
uint8_t overflows at 255 ticks (~5.1s at 20ms/tick), resetting zero_streak
to 0 and causing 2 ticks where the stale last_key is re-sent as PRESSED with
no physical button down. This produced phantom keypresses every ~5s:
- Main menu scrolled down without any button press
- Diversity Calib auto-advanced through states (last key was ENTER)
No KEY_ADC_V was printed because keypad_get_key() returned 0 throughout.

+++ Firmware/ESP32/RX5808/main/gui/lvgl_driver/lv_port_indev.c
@@ -318,7 +318,7 @@ static void keypad_read(lv_indev_drv_t * indev_drv, lv_indev_data_t * data)
     /* Release debounce: require 3 consecutive zero ADC reads before releasing.
      * This bridges ADC noise gaps on narrow-window keys (e.g. RIGHT at 2800-3000)
      * that would otherwise cause the LVGL keypad FSM to reset on every noisy tick. */
-    static uint8_t zero_streak = 0;
+    static uint32_t zero_streak = 0;  // uint32_t prevents overflow (uint8_t wraps at 255 → phantom key)
 
     uint32_t act_key = keypad_get_key();
     if(act_key != 0) {
